qqMusicProject/requirements.txt
attr==0.3.2
Brotli==1.1.0
brotlicffi==1.1.0.0
chardet==5.2.0
ConfigParser==7.2.0
cryptography==46.0.3
Cython==3.1.6
dl==0.1.0
docutils==0.22.2
filelock==3.20.0
h2==4.3.0
HTMLParser==0.0.2
importlib_resources==6.5.2
ini2toml==0.15
ipython==8.12.3
ipywidgets==8.1.7
jnius==1.1.0
js==1.0
keyring==25.6.0
mod==0.3.0
path.py==12.5.0
pkg1==0.0.3
protobuf==6.33.0
PyExecJS==1.5.1
pyodide==0.0.2
pyOpenSSL==25.3.0
pytest==8.4.2
redis==7.0.1
simplejson==3.20.2
Sphinx==8.2.3
thread==2.0.5
trove_classifiers==2025.9.11.17
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
xx==3.3.2
zstandard==0.25.0
